+++ include/roq/mbp_update.hpp
@@ -15,15 +15,18 @@
 #include "roq/string_types.hpp"
 #include "roq/uuid.hpp"
 
+#include "roq/update_action.hpp"
+
 namespace roq {
 
 //! Represents the update status of a single aggregate price level in the order book
 struct ROQ_PUBLIC MBPUpdate final {
-  double price = NaN;              //!< Price level
-  double quantity = {};            //!< Total quantity available at price
-  double implied_quantity = NaN;   //!< Total implied quantity at price (optional)
-  uint32_t price_level = {};       //!< Level of price (optional, 1-based indexing)
-  uint32_t number_of_orders = {};  //!< Number of orders at price (optional)
+  double price = NaN;               //!< Price level
+  double quantity = {};             //!< Total quantity available at price
+  double implied_quantity = NaN;    //!< Total implied quantity at price (optional)
+  uint32_t price_level = {};        //!< Level of price (0-based indexing)
+  UpdateAction update_action = {};  //!< Type of update action
+  uint16_t number_of_orders = {};   //!< Number of orders at price (optional)
 };
 
 }  // namespace roq
@@ -44,12 +47,14 @@ struct fmt::formatter<roq::MBPUpdate> {
         R"(quantity={}, )"
         R"(implied_quantity={}, )"
         R"(price_level={}, )"
+        R"(update_action={}, )"
         R"(number_of_orders={})"
         R"(}})"sv,
         value.price,
         value.quantity,
         value.implied_quantity,
         value.price_level,
+        value.update_action,
         value.number_of_orders);
   }
 };