+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes will be documented in this file.
 
 ## Head
 
+### Changed
+
+* `MBPUpdate` now supports `UpdateAction` (#236)
+
 ## 0.8.5 &ndash; 2022-06-06
 
 ### Added